refactor: ディレクトリを整理する
